keyframes and refractor-ish
