+++ src/components/userpanel/UserNav.tsx
@@ -4,27 +4,17 @@ import { useDispatch, useSelector } from "react-redux";
 import { useEthers, useLookupAddress } from "@usedapp/core";
 
 import JumboStampSystem from "./JumboStampSystem";
+import UserProfile from "./UserProfile";
 
 import UserNavIcon, { NavIconType } from "./UserNavIcon";
 import UserNavSocial, { SocialIconType } from "./UserNavSocial";
 import { selectShowingNav, setShowingNav } from "../../state/uiSlice";
 
 import logo from "../../assets/userpanel/logo.png";
-import meepocoin from "../../assets/userpanel/meepocoin.png";
-import star from "../../assets/userpanel/stampstar.png";
-
 import map from "../../assets/userpanel/buttonmap.png";
 import lobby from "../../assets/userpanel/buttonlobby.png";
 import lore from "../../assets/userpanel/buttonlore.png";
 import news from "../../assets/userpanel/buttonnews.png";
-
-import opensea from "../../assets/svgs/opensea.svg";
-import twitter from "../../assets/svgs/twitter.svg";
-import home from "../../assets/userpanel/homewebsiteicon.png";
-
-// interface NavProps {
-//   nav?: boolean;
-// }
 
 const navIcons: NavIconType[] = [
   {
@@ -45,25 +35,8 @@ const navIcons: NavIconType[] = [
   },
 ];
 
-const socialIcons: SocialIconType[] = [
-  {
-    image: twitter,
-    name: "Twitter",
-  },
-  {
-    image: opensea,
-    name: "OpenSea",
-  },
-  {
-    image: home,
-    name: "Homepage",
-  },
-];
-
 const NavContainer = styled.div`
   width: 100%;
-  /* display: flex;
-  align-items: flex-start; */
   z-index: 2;
 `;
 
@@ -153,27 +126,16 @@ const Stripes = styled.div`
   }
 `;
 
-// const NavBackgroundShape = styled.div`
-//   background-color: #dfe7f4;
-//   height: 10%;
-//   width: 10%;
-//   position: absolute;
-// `;
-
-const NavStripes = styled.div``;
-
 const NavLogo = styled.img`
   position: relative;
   padding-left: 2rem;
   margin-right: 2rem;
   top: 2rem;
-  cursor: pointer;
   width: 12%;
 `;
 
 const NavIconContainer = styled.div`
-  z-index: 100;
-  /* position: relative; */
+  z-index: 1;
   top: 3rem;
   display: inline-flex;
   position: absolute;
@@ -188,210 +150,12 @@ const NavWalletRewardsContainer = styled.div`
   justify-content: center;
 `;
 
-const NavUserContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  position: absolute;
-  top: 0;
-`;
-
-const NavUserWalletContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-right: 2rem;
-  /* background-color: black; */
-  align-items: flex-end;
-`;
-
-const NavUsernameContainer = styled.div`
-  position: relative;
-  top: 3rem;
-  right: 2rem;
-  background-color: #8ca3e7;
-  clip-path: var(--notched-sm);
-  padding: 1rem 3rem 1rem 1rem;
-  text-align: left;
-  width: fit-content;
-  height: 4rem;
-  overflow: visible;
-  margin-left: auto;
-  z-index: 1;
-  opacity: ${(props: NavProps) => (props.active ? "1" : "0")};
-  transition: all ease 0.3s;
-`;
-
-const PreNavUsername = styled.div`
-  font-size: 2.5rem;
-  color: white;
-  text-shadow: -2px 2px 0 #000, 1px 1px 0 #000, 1px -1px 0 #000,
-    -1px -1px 0 #000;
-  opacity: ${(props: NavProps) => (props.active ? "1" : "0")};
-  transition: all ease 0.3s;
-`;
-
-const PostNavUsername = styled.div`
-  font-size: 2.5rem;
-  color: white;
-  text-shadow: -2px 2px 0 #000, 1px 1px 0 #000, 1px -1px 0 #000,
-    -1px -1px 0 #000;
-  transition: all ease 0.3s;
-`;
-
-const PreNavWallet = styled.button`
-  width: fit-content;
-  cursor: pointer;
-  background-color: #fff4cb;
-  clip-path: var(--notched-sm);
-  padding: 4rem 4rem 1rem 2rem;
-  transform: ${(props: NavProps) =>
-    props.active ? "translate(10%, 0px)" : "translate(60%, 0)"};
-  transition: all ease 0.3s;
-`;
-
-const PreNavWalletText = styled.p`
-  transition: all ease 0.3s;
-  font-size: 2rem;
-  opacity: ${(props: NavProps) => (props.active ? "1" : "0")};
-`;
-
-const PostNavWallet = styled.button`
-  cursor: pointer;
-  background-color: #fff4cb;
-  clip-path: var(--notched-sm);
-  padding: 4rem 4rem 1rem 2rem;
-  transform: translate(10%, 0px);
-  transition: all ease 0.3s;
-`;
-
-const PostNavWalletText = styled.p`
-  transition: all ease 0.3s;
-  font-size: 2rem;
-  /* opacity: ${(props: NavProps) => (props.active ? "1" : "0")}; */
-`;
-
-const MeeposCollected = styled.div`
-  clip-path: var(--notched-sm);
-  width: 100%;
-  display: flex;
-  background-color: #619ee2;
-  justify-content: space-around;
-  margin-top: 1rem;
-  padding: 1rem 1rem 0.25rem 1rem;
-`;
-
-const MeeposCollectedText = styled.p`
-  font-size: 3rem;
-  color: white;
-  flex: 1 0;
-`;
-
-const MeeposCollectedStar = styled.img`
-  position: relative;
-  bottom: 0.5rem;
-  width: 25px;
-`;
-
-const StampsCollected = styled.div`
-  clip-path: var(--notched-sm);
-  width: 100%;
-  display: flex;
-  background-color: #ffcb5a;
-  justify-content: space-around;
-  margin-top: 1rem;
-  padding: 1rem 1rem 0.25rem 1rem;
-`;
-
-const StampsCollectedText = styled.p`
-  font-size: 3rem;
-  color: white;
-  flex: 1 0;
-`;
-
-const StampsCollectedStar = styled.img`
-  position: relative;
-  bottom: 0.5rem;
-  width: 25px;
-`;
-
-const DecorHorizontalContainer = styled.div`
-  display: flex;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-`;
-
-const DecorHorizontalDots = styled.span`
-  width: 1.5rem;
-  height: 1.5rem;
-  border-radius: 30px;
-  margin-right: 1rem;
-  background-color: #b2bcc3;
-`;
-
-const DecorHorizontalDots2 = styled.span`
-  width: 1.5rem;
-  height: 1.5rem;
-  border-radius: 30px;
-  margin-right: 1rem;
-  background-color: #b2bcc3;
-`;
-const DecorHorizontalDots3 = styled.span`
-  width: 100%;
-  height: 1.5rem;
-  border-radius: 30px;
-  background-color: #b2bcc3;
-`;
-
-const NavUserPfpContainer = styled.div`
-  /* position: relative;
-  top: 2rem; */
-  cursor: pointer;
-  margin-top: 2rem;
-  clip-path: var(--notched-md);
-  background-color: #393939;
-  padding: 0.25rem 0.25rem 1.25rem 0.25rem;
-`;
-
-const NavUserPfp = styled.img`
-  clip-path: var(--notched-md);
-  width: 150px;
-  height: 150px;
-`;
-
-const NavUserStats = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const NavUserSocialsContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const NavUserSocialShadow = styled.div`
-  clip-path: var(--notched-sm);
-  background-color: #393939;
-  width: fit-content;
-  padding: 0.25rem 0.25rem 0.5rem 0.25rem;
-`;
-
-const NavUserSocial = styled.img`
-  width: fit-content;
-  clip-path: var(--notched-sm);
-`;
-
 interface NavProps {
   active?: boolean;
   account?: any;
 }
 
 const UserNav = () => {
-  const [hoverActive, setHoverActive] = useState(false);
-
-  const { activateBrowserWallet, account } = useEthers();
-  const { ens } = useLookupAddress(account);
-
-  const [navActive, setActive] = useState(false);
-
   return (
     <NavContainer>
       <Triangle />
@@ -400,7 +164,6 @@ const UserNav = () => {
           <Stripes />
         </NavBackgroundDecor>
       </NavBackground>
-      {/* <NavBackgroundShape /> */}
 
       <NavLogo src={logo} />
 
@@ -411,84 +174,7 @@ const UserNav = () => {
       </NavIconContainer>
 
       <NavWalletRewardsContainer>
-        <NavUserContainer>
-          <NavUserWalletContainer>
-            <NavUsernameContainer active={hoverActive}>
-              {!ens && <PostNavUsername>A:\Welcome</PostNavUsername>}
-              {/* {!ens && <PreNavUsername active={hoverActive} />} */}
-              {ens && (
-                <PostNavUsername>{`A:\\ ${ens.substr(
-                  0,
-                  ens.length - 4
-                )}`}</PostNavUsername>
-              )}
-            </NavUsernameContainer>
-
-            {!account && (
-              <PreNavWallet
-                onMouseEnter={() => setHoverActive(true)}
-                onMouseLeave={() => setHoverActive(false)}
-                active={hoverActive}
-                onClick={() => activateBrowserWallet()}
-              >
-                {/* 0x1205...2aF4D */}
-                <PreNavWalletText active={hoverActive}>
-                  Connect Wallet
-                </PreNavWalletText>
-              </PreNavWallet>
-            )}
-
-            {account && (
-              <PostNavWallet>
-                <PostNavWalletText>
-                  {`${account.substr(0, 6)}\u2026${account.substr(
-                    account.length - 5
-                  )}`}{" "}
-                </PostNavWalletText>
-              </PostNavWallet>
-            )}
-
-            <JumboStampSystem />
-          </NavUserWalletContainer>
-
-          <NavUserStats>
-            <NavUserPfpContainer>
-              {!account && (
-                <NavUserPfp
-                  onMouseEnter={() => setHoverActive(true)}
-                  onMouseLeave={() => setHoverActive(false)}
-                  onClick={() => activateBrowserWallet()}
-                  src="https://via.placeholder.com/175x125"
-                />
-              )}
-              {account && (
-                <NavUserPfp src="https://i.seadn.io/gae/R6xFSwTpGgo7JkoVa0Acvy3EGQqdTTh5uvT74BS9NHGsMYSeknz6iFljNHC6gGyqmK_laKlkUkRhcN43mJ_OLz4SdiW5yhEsmPFhhg?auto=format&w=1000" />
-              )}
-            </NavUserPfpContainer>
-
-            <MeeposCollected>
-              <MeeposCollectedText>20485</MeeposCollectedText>
-              <MeeposCollectedStar src={meepocoin} />
-            </MeeposCollected>
-
-            <StampsCollected>
-              <StampsCollectedText>4/12</StampsCollectedText>
-              <StampsCollectedStar src={star} />
-            </StampsCollected>
-
-            <DecorHorizontalContainer>
-              <DecorHorizontalDots />
-              <DecorHorizontalDots2 />
-              <DecorHorizontalDots3 />
-            </DecorHorizontalContainer>
-
-            <NavUserSocialsContainer>
-              {socialIcons.map((socialIcon: SocialIconType) => (
-                <UserNavSocial key={socialIcon.name} socialIcon={socialIcon} />
-              ))}
-            </NavUserSocialsContainer>
-          </NavUserStats>
-        </NavUserContainer>
+        <UserProfile />
       </NavWalletRewardsContainer>
     </NavContainer>
   );
